feat: md content move mock/clock
